Working unlocked levels in stage select without relaunching. Changed sprite graphics. Added sprite left and right switching. Side collision with breakable. Changed color of breakable. Saws now working. Disabled Esc for quitting in the menus

+++ GoToHell/Saw.pde
@@ -13,7 +13,7 @@ class Saw {
    ellipse(xcor, ycor, rad * 2, rad * 2);
  }
  
- boolean insideSaw(int x, int y) {
+ boolean insideSaw(float x, float y) {
    return (double)rad - 3 >= Math.sqrt((double)(Math.pow(x - xcor, 2) + Math.pow(y - ycor, 2))) ||
    (double)rad - 3 >= Math.sqrt((double)(Math.pow(x + 20 - xcor, 2) + Math.pow(y - ycor, 2))) ||
    (double)rad - 3 >= Math.sqrt((double)(Math.pow(x - xcor, 2) + Math.pow(y + 30 - ycor, 2))) ||
